Added logic to cancelling fileChooser (#35)
Before there was an error whenever the fileChooser window was open and the user clicked cancel. I changed the fileDialogBox() function from a void to a boolean to indicate whether the user had pressed save or cancel. If they press cancel then it returns false and the confirmation window is not created. Otherwise if the user presses save then it functions as before